+++ engines/tahi_standard_tasks/app/serializers/tahi_standard_tasks/related_articles_task_serializer.rb
@@ -1,0 +1,4 @@
+module TahiStandardTasks
+  class RelatedArticlesTaskSerializer < ::TaskSerializer
+  end
+end