+++ engines/tahi_standard_tasks/spec/features/register_decision_spec.rb
@@ -2,7 +2,7 @@ require 'rails_helper'
 
 feature "Register Decision", js: true do
   let(:user) { FactoryGirl.create(:user) }
-  let(:paper) { FactoryGirl.create(:paper, :with_integration_journal, :submitted, creator: user) }
+  let(:paper) { FactoryGirl.create(:paper, :with_integration_journal, :submitted) }
   let(:task) { FactoryGirl.create(:register_decision_task, paper: paper) }
   let(:dashboard_page) { DashboardPage.new }
   let(:manuscript_page) { dashboard_page.view_submitted_paper paper }
@@ -15,56 +15,36 @@ feature "Register Decision", js: true do
   end
 
   context "Registering a decision on a paper" do
-    context "with a non-submitted Paper" do
-      before do
-        paper.update_attributes(publishing_state: "unsubmitted")
-      end
-
+    context "with a submitted Paper" do
       scenario "Participant registers a decision on the paper" do
         overlay = Page.view_task_overlay(paper, task)
-        expect(overlay.invalid_state_message).to be true
-        expect(overlay).to be_disabled
+        overlay.register_decision = "Accept"
+        overlay.decision_letter = "Accepting this because I can"
+        overlay.click_send_email_button
+        wait_for_ajax
+        expect(task.reload.completed?).to be true
       end
-    end
 
-    scenario "Participant registers a decision on the paper" do
-      overlay = Page.view_task_overlay(paper, task)
-      overlay.register_decision = "Accept"
-      overlay.decision_letter = "Accepting this because I can"
-      overlay.click_send_email_button
-      expect(overlay).to be_completed
-    end
-
-    scenario "Disable inputs upon card completion" do
-      overlay = Page.view_task_overlay(paper, task)
-      overlay.register_decision = "Accept"
-      overlay.decision_letter = "Accepting this because I can"
-      overlay.click_send_email_button
-      expect(overlay).to be_completed
-      expect(overlay.success_state_message).to be true
-      expect(first('input[name=decision]')).to be_disabled
-    end
-
-    scenario "persist the decision radio button" do
-      overlay = Page.view_task_overlay(paper, task)
-      overlay.register_decision = "Reject"
-      wait_for_ajax
-      overlay.radio_selected?
-
-      visit current_path # Revisit
-      overlay = Page.view_task_overlay(paper, task)
-      expect(find("input[value='reject']")).to be_checked
-    end
-
-    context "with a non submitted Paper" do
-      scenario "display flash message and disable card" do
+      scenario "Disable inputs upon card completion" do
         overlay = Page.view_task_overlay(paper, task)
+        overlay.register_decision = "Accept"
+        overlay.decision_letter = "Accepting this because I can"
+        overlay.click_send_email_button
+        wait_for_ajax
+        expect(task.reload.completed?).to be true
+        expect(overlay.success_state_message).to be true
+        expect(first('input[name=decision]')).to be_disabled
+      end
+
+      scenario "persist the decision radio button" do
+        overlay = Page.view_task_overlay(paper, task)
+
         overlay.register_decision = "Reject"
         wait_for_ajax
         overlay.radio_selected?
 
         visit current_path # Revisit
-        overlay = Page.view_task_overlay(paper, task)
+        wait_for_ajax
         expect(find("input[value='reject']")).to be_checked
       end
     end
@@ -85,6 +65,19 @@ feature "Register Decision", js: true do
       overlay.find("#accordion h4.panel-title a").click # open Accordion
       expect(overlay.previous_decisions.first.letter).to eq("Please revise the manuscript. After line break")
       expect(overlay.previous_decisions.first.letter).to_not include "<br>"
+    end
+  end
+
+  context "with an unsubmitted Paper" do
+    before do
+      paper.update_attributes!(publishing_state: 'unsubmitted')
+      paper.reload
+    end
+
+    scenario "Participant cannot register a decision on the paper" do
+      overlay = Page.view_task_overlay(paper, task)
+      expect(overlay.invalid_state_message).to be true
+      expect(overlay).to be_disabled
     end
   end
 end